refactor: reestructurar controladores, modelos y rutas en subcarpetas según su entidad
