+++ models/sampleModel/sampleModel.js
@@ -1,4 +1,4 @@
-import pool from "../config/db.js";
+import pool from "../../config/db.js";
 
 class Muestra {
   static async findAll() {
