+++ routes/userRoutes/userRoutes.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import UserController from '../controllers/userController.js';
+import UserController from '../../controllers/userController/userController.js';
 
 const router = express.Router();
